the code in the main app file has been abstracted for an easier developer onboarding experience
